requirements.txt
pyfiglet==0.8.post1
python_dateutil==2.8.1
python_nmap==0.7.1
